+++ server_code/authorisation.py
@@ -5,10 +5,32 @@
 #
 # This software is published at https://github.com/anvilistas/anvil-extras
 import functools
+from operator import itemgetter
 
 import anvil.users
+from anvil.tables import app_tables
 
 __version__ = "2.6.1"
+
+config = {"get_roles": itemgetter("roles")}
+
+
+def set_config(**kwargs):
+    if "get_roles" in kwargs:
+        _set_user_roles_getter(kwargs["get_roles"])
+
+
+def _get_roles_from_table(table_name, user):
+    return getattr(app_tables, table_name).get(user=user)["roles"]
+
+
+def _set_user_roles_getter(option):
+    if option is None:
+        config["get_roles"] = itemgetter("roles")
+    elif isinstance(option, str):  # table name
+        config["get_roles"] = functools.partial(_get_roles_from_table, option)
+    else:
+        raise TypeError("get_roles: option is not valid.")
 
 
 def authentication_required(func):
@@ -38,7 +60,7 @@ def has_permission(permissions):
     try:
         user_permissions = set(
             permission["name"]
-            for role in user["roles"]
+            for role in config["get_roles"](user)
             for permission in role["permissions"]
         )
     except TypeError: