+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Unreleased
 
+## Enhancements
+* authorisation - added config option to avoid adding a roles column to the users table
+  https://github.com/anvilistas/anvil-extras/pull/516
+
 ## Bug Fixes
 * persistence - fix get method
   https://github.com/anvilistas/anvil-extras/issues/523
+
 
 # v2.6.1 27-Mar-2024
 
@@ -21,6 +26,8 @@
 ## Minor Changes
 * designer hints - add some designer hints to components
   https://github.com/anvilistas/anvil-extras/pull/512
+* authorisation - added config option to avoid adding a roles column to the users table
+  https://github.com/anvilistas/anvil-extras/pull/516
 
 # v2.6.0 09-Feb-2024